[fix] WS2PStreamer moving was not complete
