+++ app/modules/ws2p/lib/WS2PClient.ts
@@ -1,9 +1,9 @@
 import {Server} from "../../../../server"
 import {WS2PConnection, WS2PPubkeyLocalAuth, WS2PPubkeyRemoteAuth} from "./WS2PConnection"
-import {WS2PStreamer} from "../../../lib/streams/WS2PStreamer"
 import {Key} from "../../../lib/common-libs/crypto/keyring"
 import {WS2PMessageHandler} from "./impl/WS2PMessageHandler"
 import {WS2PConstants} from "./constants"
+import { WS2PStreamer } from "./WS2PStreamer";
 
 export class WS2PClient {
